backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Kubernetes client
kubernetes==28.1.0

# Authentication and security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# Rate limiting
slowapi==0.1.9

# CORS
python-dotenv==1.0.0

# Logging and monitoring
python-json-logger==2.0.7
